refactor Section.js
