+++ components/home-page/index.ts
@@ -1,5 +1,5 @@
 import { Header, HeaderContent } from './Header';
 import { Nav, NavLink } from './Nav';
-import Section from './Section';
+import { Section } from './Section';
 
 export { Header, HeaderContent, Nav, NavLink, Section };
